% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylocompmeth-package.R
\docType{package}
\name{phylocompmeth-package}
\alias{phylocompmeth}
\alias{phylocompmeth-package}
\title{phylocompmeth: Phylogenetic Comparative Methods}
\description{
The function included in this package
}
\author{
\strong{Maintainer}: Avery Shepherd \email{ashepherd@utexas.edu}

}
\keyword{internal}
