% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nexus.R
\name{read.nexus}
\alias{read.nexus}
\title{An interactive power plotting function}
\usage{
read.nexus(file, tree.names = NULL, force.multi = FALSE)
}
\arguments{
\item{file}{a file name specified by either a variable of mode character, or a double-quoted string.}

\item{tree.names}{if there are several trees to be read, a vector of mode character giving names to the individual trees (by default, this uses the labels in the NEXUS file if these are present).}

\item{force.multi}{a logical value; if TRUE, an object of class "multiPhylo" is always returned even if the file contains a single tree (see details).}
}
\description{
This function allows you explore the relationship between sample size, alpha level, mean, standard deviation, and power
}
\keyword{nexus}
\keyword{read}
